CSV (LoggerCSV) Analytics / statistics (#496)
* Prototypes

* CSV analytics first iteration

* Reduce complexity

* m

* Refactoring

* Refactoring file logging
