+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
-    "setuptools>=62.6,<66",
-    "setuptools-scm>=6.2,<8",
+    "setuptools",
+    "setuptools-scm",
     "wheel",
 ]
