+++ src/daf/logging/__init__.py
@@ -1,4 +1,4 @@
 "Package used for logging"
-from ._logging import *
 from .sql import *
 from .tracing import *
+from ._logging import *
